refactor(scripts): extract transaction send helpers into transactions.ts
